fix(P0): FX-13 migrar Double a BigDecimal + Flyway

+++ fleet-management/src/main/java/com/fleet/management/dto/recorrido/RecorridoRequest.java
@@ -38,5 +38,6 @@ public class RecorridoRequest {
     private Long tarjetaCombustibleId;
 
     @Positive(message = "El importe abastecido debe ser mayor a cero")
-    private Double importeAbastecido;
+    // FX-13: migrado de Double a BigDecimal para precision monetaria.
+    private BigDecimal importeAbastecido;
 }
